backend/requirements.txt
# Django and REST Framework
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.0
django-filter==24.1

# Database
psycopg2-binary==2.9.10
pgvector==0.2.5

# Task Queue and Caching
celery==5.5.3
redis==6.2.0
django-redis==5.4.0

# Environment and Utilities
python-dotenv==1.1.1
requests==2.31.0

# Image Processing
Pillow==10.4.0

# Document Processing
PyMuPDF==1.23.26
python-multipart==0.0.9

# Production Server
gunicorn==21.2.0

# Health Checks
curl_cffi==0.6.0

# System Monitoring
psutil==5.9.8

# AI/ML Dependencies
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.36.2
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4
ollama==0.1.7

# Additional AI utilities
huggingface-hub==0.19.4
tokenizers==0.15.0 